Add Makefiles to examples (#596)
Building the examples presents a source of friction to new
users who want to experiment with the language. This is
especially true when an example needs to be compiled with
the standard library. This commit adds a standard Makefile
for examples to make them easy to build and provides a
template for adding new examples.